Fix non-navigatable download links
Instead of clicking on the, assuming we will get
to some renderable binary page (like PDF reader),
we pre-detect if the page is a known webpage extension.
If it's not, we download the file using straight
HTTP requests. This could fail if the download requires
HTTP headers/etc, but this is rare enough that if
this becomes a problem, we can add an option and deal
with it then, as this is a complex problem.