Implement infinite 'next page' loop detection
This introduces a chance to avoid infinite 'next page' loops
where sites do not disallow clicking the next button, even though
we're on the final page (continuing to show the last result page).
This is clearly really bad web design, but it's out there.

If the previous past five result pages are identical, AutoScrape
will now detect this and break out of the 'next page' loop.