Asynchronous document readyState waiting doesn't work
Instead of trying to get the onreadystatechange callback
strategy to work when deciding if a page has loaded, we're just going to rely
on the normal document.readyState and then --form-submit-wait flag
for controlling timeouts.

The only hole this leaves is if people want to do click-only crawls
inside of single page apps, this does not have a good way to
properly time those.