1.6.12 Performance improvement, stability fixes
This is a major memory and general performance improvement,
accomplished by loading link and button vectors upon their
use. If the scraper doesn't try to use a link or button, we
never bother loading the vectors. This helps drastically on
results pages with thousands of URLs, where we just want to
look for a button, for example.

Some other improvements/changes:

- Selenium backend: better async page load detection
- Selenium backend: don't use 'go' as default submit text
- Selenium backend: no default next button text
- Requests backend: include retry wait to 30s
- Requests backend: include common browser user-agent
- Remove stub "ML" backend